+++ src/js/components/lib/cells/Door.js
