+++ src/js/components/lib/cells/Wall.js
